+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Not released:
 
 * Warn on undefined return value of `dragBoundFunc`.
+* Better calculations for `container.getClientRect()`
 
 ## 4.0.15 - 2019-10-15